added navigation-types.ts
